# bff/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Basalt Arena BFF API
  version: 1.0.0
  description: REST API used by the public arena client and the custom admin panel.
servers:
  - url: http://localhost:3001
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Login with email or handle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Auth payload with access and refresh tokens
  /api/v1/auth/register:
    post:
      tags: [auth]
      summary: Register a member account
      description: If DEV_REGISTER_KEY is configured, x-dev-register-key is required.
      parameters:
        - $ref: '#/components/parameters/DevRegisterKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registered
        '403':
          description: Registration is restricted
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Rotate refresh token and issue a new token pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              additionalProperties: false
              properties:
                refreshToken:
                  type: string
                  minLength: 10
      responses:
        '200':
          description: New auth payload
        '401':
          description: Invalid, expired, or already used refresh token
  /api/v1/auth/logout:
    post:
      tags: [auth]
      summary: Revoke current session
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Logged out
...
  /api/v1/sprints:
    get:
      tags: [sprints]
      summary: List published, non-archived sprints with solutions
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/SortBy'
      responses:
        '200':
          description: List of sprints
  /api/v1/sprints/{id}:
    get:
      tags: [sprints]
      summary: Sprint detail with solutions
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/SortBy'
      responses:
        '200':
          description: Sprint detail
        '404':
          description: Sprint not found or not public
...
  /api/v1/sprints/{id}/submissions:
    post:
      tags: [sprints]
      summary: Submit or update current user's solution for a sprint
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmissionUpsertRequest'
      responses:
        '201':
          description: Created or updated submission
  /api/v1/submissions:
    post:
      tags: [sprints]
      summary: Submit or update current user's solution for the active sprint
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmissionUpsertRequest'
      responses:
        '201':
          description: Created or updated submission
...
  /api/v1/admin/audit-logs:
    get:
      tags: [admin]
      summary: List audit logs
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Skip'
        - $ref: '#/components/parameters/Take'
        - in: query
          name: userId
          schema: { type: string }
      responses:
        '200':
          description: Paginated audit logs
  /api/v1/admin/users:
    get:
      tags: [admin]
      summary: List users
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: search
          schema: { type: string, maxLength: 200 }
        - $ref: '#/components/parameters/Skip'
        - $ref: '#/components/parameters/Take'
      responses:
        '200':
          description: Paginated users
  /api/v1/admin/users/{id}:
    patch:
      tags: [admin]
      summary: Update user account and profile fields
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPatchUserRequest'
      responses:
        '200':
          description: Updated user
  /api/v1/admin/users/{userId}/achievements:
    get:
      tags: [admin]
      summary: List achievements granted to a user
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User achievements
  /api/v1/admin/users/{userId}/achievements/{achievementId}:
    post:
      tags: [admin]
      summary: Grant achievement to user
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/AchievementId'
      responses:
        '201':
          description: Granted
    delete:
      tags: [admin]
      summary: Revoke achievement from user
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/AchievementId'
      responses:
        '200':
          description: Revoked
  /api/v1/admin/users/{userId}/sprints:
    get:
      tags: [admin]
      summary: List sprint access and submission activity for a user
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User sprint activity
  /api/v1/admin/sprints:
    get:
      tags: [admin]
      summary: List all sprints, including drafts and archived records
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Sprints
    post:
      tags: [admin]
      summary: Create sprint
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSprintWriteRequest'
      responses:
        '201':
          description: Created sprint
  /api/v1/admin/sprints/{id}:
    patch:
      tags: [admin]
      summary: Update sprint fields
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSprintPatchRequest'
      responses:
        '200':
          description: Updated sprint
  /api/v1/admin/sprints/{id}/activate:
    post:
      tags: [admin]
      summary: Mark sprint active and deactivate all others
      description: >
        Sets startsAt to UTC start of the current calendar day. If endsAt is missing,
        in the past, or not after that new start, sets endsAt to +7 days from the new start;
        otherwise keeps the existing future endsAt (manual deadline).
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Activated sprint
  /api/v1/admin/sprints/{id}/duplicate:
    post:
      tags: [admin]
      summary: Duplicate sprint as an inactive public draft copy
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '201':
          description: Duplicated sprint
  /api/v1/admin/sprints/{sprintId}/access:
    get:
      tags: [admin]
      summary: List explicit sprint access rows
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/SprintId'
      responses:
        '200':
          description: Access rows
    put:
      tags: [admin]
      summary: Upsert one user's sprint access
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/SprintId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSprintAccessRequest'
      responses:
        '200':
          description: Access row
  /api/v1/admin/sprints/{sprintId}/access/batch:
    put:
      tags: [admin]
      summary: Upsert sprint access for multiple users
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/SprintId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminBatchSprintAccessRequest'
      responses:
        '200':
          description: Batch result
  /api/v1/admin/sprints/{sprintId}/access/{userId}:
    delete:
      tags: [admin]
      summary: Delete explicit sprint access for a user
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/SprintId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Deleted
  /api/v1/admin/sprints/{sprintId}/achievements/{achievementId}/grant-sprint:
    post:
      tags: [admin]
      summary: Grant achievement to sprint participants
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/SprintId'
        - $ref: '#/components/parameters/AchievementId'
      responses:
        '200':
          description: Grant result
  /api/v1/admin/submissions:
    get:
      tags: [admin]
      summary: List submissions
      description: Only the ADMIN role may access this admin area.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: sprintId
          schema: { type: string }
        - in: query
          name: userId
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, example: 'PENDING,REVIEWED' }
          description: Comma-separated SubmissionStatus values.
        - $ref: '#/components/parameters/Skip'
        - $ref: '#/components/parameters/Take'
      responses:
        '200':
          description: Paginated submissions with user and sprint
  /api/v1/admin/submissions/{id}:
    patch:
      tags: [admin]
      summary: Update submission review fields
      description: Only the ADMIN role may access this admin area.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPatchSubmissionRequest'
      responses:
        '200':
          description: Updated submission row
  /api/v1/admin/submissions/batch-accept:
    post:
      tags: [admin]
      summary: Accept multiple submissions and assign mentorScore 100
      description: Only the ADMIN role may access this admin area.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              additionalProperties: false
              properties:
                ids:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items: { type: string }
      responses:
        '200':
          description: Batch result
  /api/v1/admin/achievements:
    get:
      tags: [admin]
      summary: List achievements
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Achievements
    put:
      tags: [admin]
      summary: Create or update achievement
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpsertAchievementRequest'
      responses:
        '200':
          description: Achievement
  /api/v1/admin/achievements/{id}:
    delete:
      tags: [admin]
      summary: Delete achievement
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Deleted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string }
    UserId:
      in: path
      name: userId
      required: true
      schema: { type: string }
    SprintId:
      in: path
      name: sprintId
      required: true
      schema: { type: string }
    AchievementId:
      in: path
      name: achievementId
      required: true
      schema: { type: string }
    SortBy:
      in: query
      name: sortBy
      schema:
        type: string
        enum: [efficiency, likes, mentor]
        default: efficiency
    Skip:
      in: query
      name: skip
      schema: { type: integer, minimum: 0, default: 0 }
    Take:
      in: query
      name: take
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    DevRegisterKey:
      in: header
      name: x-dev-register-key
      required: false
      schema: { type: string }
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email: { type: string, minLength: 1, maxLength: 320 }
        password: { type: string, minLength: 1, maxLength: 512 }
    RegisterRequest:
      type: object
      required: [email, handle, password]
      additionalProperties: false
      properties:
        email: { type: string, format: email, maxLength: 320 }
        handle:
          type: string
          minLength: 2
          maxLength: 64
          pattern: '^@?[a-zA-Z0-9_-]+$'
        password: { type: string, minLength: 8, maxLength: 512 }
    SubmissionUpsertRequest:
      type: object
      required: [repoUrl]
      additionalProperties: false
      properties:
        repoUrl: { type: string, format: uri, maxLength: 2048 }
        demoUrl:
          oneOf:
            - { type: string, format: uri, maxLength: 2048 }
            - { type: string, const: '' }
    AdminPatchUserRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        role:
          type: string
          enum: [ADMIN, USER]
        email: { type: string, format: email, maxLength: 320 }
        points: { type: integer, minimum: 0 }
        handle: { type: string, minLength: 2, maxLength: 64 }
        bio: { type: string, maxLength: 500 }
        telegram: { type: string, maxLength: 64 }
        github: { type: string, maxLength: 512 }
        realName: { type: string, maxLength: 120 }
        stack:
          type: array
          maxItems: 30
          items: { type: string, minLength: 1, maxLength: 64 }
        moneyEarned: { type: integer, minimum: 0, maximum: 2000000000 }
    AdminSprintWriteRequest:
      type: object
      required: [slug, title, tabLabel, completedLabel]
      additionalProperties: false
      properties:
        slug: { type: string, pattern: '^[a-z0-9-]+$', maxLength: 128 }
        title: { type: string, minLength: 1, maxLength: 256 }
        tabLabel: { type: string, minLength: 1, maxLength: 128 }
        tabIcon:
          oneOf:
            - { type: string, maxLength: 64 }
            - { type: 'null' }
        completedLabel: { type: string, minLength: 1, maxLength: 128 }
        tags:
          type: array
          maxItems: 50
          items: { type: string, maxLength: 64 }
        active: { type: boolean }
        published: { type: boolean }
        archived: { type: boolean }
        startsAt:
          oneOf:
            - { type: string, format: date-time }
            - { type: 'null' }
        endsAt:
          oneOf:
            - { type: string, format: date-time }
            - { type: 'null' }
        brief:
          type: object
          additionalProperties: true
        metrics:
          type: object
          additionalProperties: true
    AdminSprintPatchRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        slug: { type: string, pattern: '^[a-z0-9-]+$', maxLength: 128 }
        title: { type: string, minLength: 1, maxLength: 256 }
        tabLabel: { type: string, minLength: 1, maxLength: 128 }
        tabIcon:
          oneOf:
            - { type: string, maxLength: 64 }
            - { type: 'null' }
        completedLabel: { type: string, minLength: 1, maxLength: 128 }
        tags:
          type: array
          maxItems: 50
          items: { type: string, maxLength: 64 }
        active: { type: boolean }
        published: { type: boolean }
        archived: { type: boolean }
        startsAt:
          oneOf:
            - { type: string, format: date-time }
            - { type: 'null' }
        endsAt:
          oneOf:
            - { type: string, format: date-time }
            - { type: 'null' }
        brief:
          type: object
          additionalProperties: true
        metrics:
          type: object
          additionalProperties: true
    AdminSprintAccessRequest:
      type: object
      required: [userId, canSubmit, canView]
      additionalProperties: false
      properties:
        userId: { type: string }
        canSubmit: { type: boolean }
        canView: { type: boolean }
    AdminBatchSprintAccessRequest:
      type: object
      required: [userIds, canSubmit, canView]
      additionalProperties: false
      properties:
        userIds:
          type: array
          minItems: 1
          maxItems: 200
          items: { type: string }
        canSubmit: { type: boolean }
        canView: { type: boolean }
    AdminPatchSubmissionRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        mentorScore: { type: integer, minimum: 0, maximum: 100 }
        status:
          type: string
          enum: [PENDING, REVIEWED, ACCEPTED, REJECTED]
        repoUrl: { type: string, format: uri, maxLength: 2048 }
        demoUrl:
          oneOf:
            - { type: string, format: uri, maxLength: 2048 }
            - { type: 'null' }
        mentorComment:
          oneOf:
            - { type: string, maxLength: 8000 }
            - { type: 'null' }
    AdminUpsertAchievementRequest:
      type: object
      required: [slug, title, subtitle, icon]
      additionalProperties: false
      properties:
        id: { type: string }
        slug: { type: string, pattern: '^[a-z0-9-]+$', maxLength: 128 }
        title: { type: string, minLength: 1, maxLength: 256 }
        subtitle: { type: string, minLength: 1, maxLength: 512 }
        icon: { type: string, minLength: 1, maxLength: 128 }